Move models to their own package
branch:
